Move consumer to app in order to use common entities
You can now use `flask consume-kafka` command.
Use Dataset, Reuse and Organization dataclasses declared in app/domain in kafka consumer.
Add a dataclass helper (`load_from_dict`) to prevent failure on object initialization.
Update docker configuration accordingly.
