+++ Dockerfiles/Dockerfile.consumer
@@ -1,0 +1,12 @@
+FROM python:3.9
+
+WORKDIR /app
+
+# install dependencies
+COPY requirements.txt .
+RUN pip install -r requirements.txt
+
+# copy project
+COPY . .
+
+ENTRYPOINT ["flask", "consume-kafka"]
