+++ docker-compose.yml
@@ -56,7 +56,8 @@ services:
 
   kafka-consumer:
     build:
-      context: kafka-consumer
+      context: .
+      dockerfile: Dockerfiles/Dockerfile.consumer
     container_name: consumer01
     environment:
       ELASTIC_HOST: search
@@ -64,6 +65,6 @@ services:
       KAFKA_HOST: broker  # This value depends on your network configuration
       KAFKA_PORT: 29092
       KAFKA_API_VERSION: 2.5.0
-      LOGGING_LEVEL: 20
+      CONSUMER_LOGGING_LEVEL: 20
 
 # image: docker.elastic.co/elasticsearch/elasticsearch:7.14.1
